fix: NumPy 2.x unsigned integer overflow in all test files
Extended the unsigned integer overflow fix to all test files that use
np.array() or scipy.sparse.coo_matrix() with unsigned dtypes and -1 values.

NumPy 2.x raises OverflowError when creating arrays with out-of-bound values
(e.g., -1 for unsigned integer types). NumPy 1.x allowed overflow with a
deprecation warning.

Solution: Create arrays without dtype first, then use .astype(dtype) to allow
overflow wrapping (e.g., -1 → max value for unsigned types).

Fixed files:
- tests/core/test_einsum.py: test_avos, test_identity
- tests/sparse/test_sparse_matmul.py: test_avos (multiple coo_matrix calls)
- tests/avos/test_apsp.py: test_apsp_sparse_impl

TEST RESULTS - ALL TESTS PASSING:
✅ Python 3.10 + NumPy 2.2.6: 117/117 tests PASSING (100%)
✅ Python 3.11 + NumPy 2.3.4: 117/117 tests PASSING (100%)
✅ Python 3.12 + NumPy 2.3.4: 117/117 tests PASSING (100%)

NumPy 2.x migration now complete with full test coverage across all test suites!

+++ tests/core/test_einsum.py
@@ -15,16 +15,17 @@ import pytest
 ])
 def test_avos(dtype):
     # test simple avos matmul
+    # NumPy 2.x: use astype() for unsigned dtypes to allow overflow wrapping
     A = np.array([[-1,2, 3, 0, 0],
                   [0,-1, 0, 2, 0],
                   [0, 0, 1, 0, 0],
                   [0, 0, 0,-1, 0],
-                  [2, 0, 0, 0, 1]], dtype=dtype)
+                  [2, 0, 0, 0, 1]]).astype(dtype)
     S = np.array([[-1, 2, 3, 4, 0],
                   [0,-1, 0, 2, 0],
                   [0, 0, 1, 0, 0],
                   [0, 0, 0,-1, 0],
-                  [2, 4, 5, 0, 1]], dtype=dtype)
+                  [2, 4, 5, 0, 1]]).astype(dtype)
     assert_equal(einsum('ij, jk', A, A, avos=True), S)
 
     # test vector mat mul
@@ -32,7 +33,7 @@ def test_avos(dtype):
                        [0,-1, 0, 2, 0],
                        [0, 0, 1, 0, 0],
                        [0, 0, 0,-1, 0],
-                       [2, 4, 5, 8, 1]], dtype=dtype)
+                       [2, 4, 5, 8, 1]]).astype(dtype)
     u = np.array([2, 0, 0, 0, 0], dtype=dtype)
     v = np.array([0, 3, 0, 0, 0], dtype=dtype)
     u_lambda = np.array([2, 4, 5, 8, 0])
@@ -53,11 +54,12 @@ def test_avos(dtype):
     np.uint64
 ])
 def test_identity(dtype):
+    # NumPy 2.x: use astype() for unsigned dtypes to allow overflow wrapping
     A = np.array([[-1, 2, 3, 0, 0],
                   [ 0,-1, 0, 2, 0],
                   [ 0, 0, 1, 0, 0],
                   [ 0, 0, 0,-1, 0],
-                  [ 2, 0, 0, 0, 1]], dtype=dtype)
+                  [ 2, 0, 0, 0, 1]]).astype(dtype)
     I = np.array([[ 1, 0, 0, 0, 0],
                   [ 0, 1, 0, 0, 0],
                   [ 0, 0, 1, 0, 0],
